// src/modules/invoice/usecase/find-invoice/find-invoice.usecase.dto.ts
import Address from "../../../@shared/domain/value-object/address"

export interface FindInvoiceUseCaseInputDto {
    id: string
}

export interface FindInvoiceUseCaseOutputDto {
    id: string
    name: string
    document: string
    address: {
        street: string
        number: string
        complement: string
        city: string
        state: string
        zipCode: string
    }
    items: {
        id: string
        name: string
        price: number
    }[]
    total: number
    createdAt: Date
}